requirements.txt
opencv-python>=4.9.0.80
numpy>=1.24.4
ultralytics>=8.1.31
pydantic>=2.6.4
pyserial>=3.5